backend/requirements.txt
# Python Backend Requirements for Power BI Training Portal

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# GraphQL
strawberry-graphql[fastapi]==0.219.0
graphql-core==3.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyjwt==2.8.0

# Database & Azure Integration
pyodbc==5.0.1
azure-identity==1.15.0
azure-storage-blob==12.19.0

# Microsoft Fabric SDK
semantic-link-labs==0.5.0
msfabricpysdkcore==1.0.0

# Data Processing
pandas==2.1.4
pydantic==2.5.3
pydantic-settings==2.1.0

# API Documentation
fastapi-pagination==0.12.14

# CORS & Middleware
python-cors==1.0.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Video Processing (Optional)
python-magic==0.4.27

# Logging & Monitoring
loguru==0.7.2
